session4_data_models_p2
Wow I'm struggling with git lol, cant seem to pull the various branches with homework to local, just master... so making this PR is a total hack via github branch!

I've had the advantage of being able to see the answers this time round which made implementing the CRUDs easier... but I did not copy them directly! My strategy was to give it a go by directly copying the bits that were already implemented and changing the names (which is how I did it last time)... that turned out to be wrong when I checked it against the solution. What I found was that the pieces to feed into the function (are these methods?) was that the fields I needed were either in the module of entry/anchor or in the handlers... I'm not sure where I should be lifting the fields from to be honest or how they connect yet despite watching the videos a bunch of times... so behind on validation at this point...